x86/resctrl: Claim get_domain_from_cpu() for resctrl
get_domain_from_cpu() is a handy helper that both the arch code and
resctrl need to use. Rename it resctrl_get_domain_from_cpu() so it
gets moved out to /fs, and exported back to the arch code.

(cherry picked from commit ffd2335a26f1fc7f1ce70d260728885c5fda89a8 https://git.kernel.org/pub/scm/linux/kernel/git/morse/linux.git/log/?h=mpam/snapshot/v6.7-rc2)
